% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{read_excel_allsheets}
\usage{
read_excel_allsheets(filename)
}
\arguments{
\item{filename}{complete name of the excel worksheet}
}
\description{
Reads all sheets of a worksheet and save it in a list of tibbles
}
